Extracted fileLogger to Utils
