+++ core/src/main/java/org/kiwix/kiwixmobile/core/CoreApp.kt
@@ -20,10 +20,8 @@ package org.kiwix.kiwixmobile.core
 import android.app.Application
 import android.content.Context
 import android.os.Build
-import android.os.Environment
 import android.os.StrictMode
 import android.os.StrictMode.VmPolicy
-import android.util.Log
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.multidex.MultiDex
 import com.jakewharton.threetenabp.AndroidThreeTen
@@ -31,8 +29,7 @@ import org.kiwix.kiwixmobile.core.data.local.KiwixDatabase
 import org.kiwix.kiwixmobile.core.di.components.CoreComponent
 import org.kiwix.kiwixmobile.core.di.components.DaggerCoreComponent
 import org.kiwix.kiwixmobile.core.downloader.DownloadMonitor
-import java.io.File
-import java.io.IOException
+import org.kiwix.kiwixmobile.core.utils.files.FileUtils.writeLogFile
 import javax.inject.Inject
 
 abstract class CoreApp : Application() {
@@ -127,32 +124,4 @@ abstract class CoreApp : Application() {
       )
     }
   }
-
-  private fun writeLogFile() {
-    if (isExternalStorageWritable) {
-      val appDirectory =
-        File(Environment.getExternalStorageDirectory().toString() + "/Kiwix")
-      val logFile = File(appDirectory, "logcat.txt")
-      Log.d("KIWIX", "Writing all logs into [" + logFile.path + "]")
-      // create app folder
-      if (!appDirectory.exists()) {
-        appDirectory.mkdir()
-      }
-      if (logFile.exists() && logFile.isFile) {
-        logFile.delete()
-      }
-      // clear the previous logcat and then write the new one to the file
-      try {
-        logFile.createNewFile()
-        Runtime.getRuntime().exec("logcat -c")
-        Runtime.getRuntime().exec("logcat -f " + logFile.path + " -s kiwix")
-      } catch (e: IOException) {
-        Log.e("KIWIX", "Error while writing logcat.txt", e)
-      }
-    }
-  }
-
-  /* Checks if external storage is available for read and write */
-  val isExternalStorageWritable: Boolean
-    get() = Environment.MEDIA_MOUNTED == Environment.getExternalStorageState()
 }
