+++ core/src/main/java/org/kiwix/kiwixmobile/core/settings/CorePrefsFragment.java
@@ -54,6 +54,7 @@ import org.kiwix.kiwixmobile.core.utils.SharedPreferenceUtil;
 import static org.kiwix.kiwixmobile.core.utils.ConstantsKt.RESULT_RESTART;
 import static org.kiwix.kiwixmobile.core.utils.SharedPreferenceUtil.PREF_NIGHT_MODE;
 import static org.kiwix.kiwixmobile.core.utils.SharedPreferenceUtil.PREF_STORAGE;
+import static org.kiwix.kiwixmobile.core.utils.files.FileUtils.isExternalStorageWritable;
 
 public abstract class CorePrefsFragment extends PreferenceFragmentCompat implements
   SettingsContract.View,
@@ -217,7 +218,7 @@ public abstract class CorePrefsFragment extends PreferenceFragmentCompat impleme
   }
 
   private void clearAllNotes() {
-    if (CoreApp.getInstance().isExternalStorageWritable()) {
+    if (isExternalStorageWritable()) {
       if (ContextCompat.checkSelfPermission(getActivity(),
         Manifest.permission.WRITE_EXTERNAL_STORAGE)
         != PackageManager.PERMISSION_GRANTED) {
